refactor: move extra encoding utilities into helper
